First step to customizable slugs
For the primary nav items, introduce a `rewrite_id` new argument to use it as the key for the customizable slug. Update the Custom URLs admin screen to allow this customization using a specific post meta to store the customized slugs. One benefit of using a post type is we make sure the directory slugs are unique