Improve Site wide notices layout on Legacy
Use the notice template tags inside bp_message_get_notices() to output the subject and text of the notice.
Remove extra the p tag containing the title and text as the text will be wpautop thanks to the bp_message_notice_text template tag.
Adapt the Legacy styles to the new markup.